handle overwrite correctly

+++ scripts/train-class1-allele-specific-models.py
@@ -70,10 +70,10 @@ parser.add_argument(
 if __name__ == "__main__":
     args = parser.parse_args()
 
-    if exists(args.output_dir):
-        if args.overwrite:
-            rmtree(args.output_dir)
-    else:
+    if exists(args.output_dir) and args.overwrite:
+        rmtree(args.output_dir)
+        makedirs(args.output_dir)
+    elif not exists(args.output_dir):
         makedirs(args.output_dir)
     allele_groups, _ = load_data(
         args.binding_data_csv_path,
